fix samba for early genus abort
If the coefficient domain is not a field, the genus criterion
(genus=number of gaps) does not hold, because the basis might be
missing an element of higher grade, but smaller euclideanSize in its
leading coefficient. So we postpone the application of the genus
criterion until we have found a leading coefficient that is a unit.

DANGER: It is not yet clear whether this is enough. It might happen
that we have added an element of grade n and a unit as leading
coefficient but, if the computation continues, there would be another
element of smaller grade but slightly bigger euclideanSize added to
the basis. In other words, for a Euclidean domain the basis computed
with genus criterion might not be the true basis that reduces every
element to zero.