refactor to use MOD as parameter
Functions corresponding to the modular group are abstracted out
and put into their own domain. Instead of CongruenceSubgroupGamma0 and
CongruenceSubgroupGamma1, we now use QEtaConditionsModularGamma0 and
QEtaConditionsModularGamma1 as parameters for domains.

Additionally, we unified SymbolicEtaQuotient and
SymbolicGeneralizedEtaQuotient. When we speak of eta-quotient we can
mean a generalized eta-quotient.

Remove QEtaCofactorSpace and QEtaAuxiliaryFunctions as well as
the (already unused) code to annihilate
coefficients. (QAuxiliaryModularPackage).

The tests and macros have been adapted accordingly.